refactor secrets and settings setup
