+++ app/models/payment_providers/base_provider.rb
@@ -3,6 +3,8 @@
 module PaymentProviders
   class BaseProvider < ApplicationRecord
     include PaperTrailTraceable
+    include SecretsStorable
+    include SettingsStorable
 
     self.table_name = 'payment_providers'
 
@@ -16,31 +18,8 @@ module PaymentProviders
     has_many :payments, dependent: :nullify, foreign_key: :payment_provider_id
     has_many :refunds, dependent: :nullify, foreign_key: :payment_provider_id
 
-    encrypts :secrets
-
     validates :code, uniqueness: { scope: :organization_id }
     validates :name, presence: true
-
-    def secrets_json
-      JSON.parse(secrets || '{}')
-    end
-
-    def push_to_secrets(key:, value:)
-      self.secrets = secrets_json.merge(key => value).to_json
-    end
-
-    def get_from_secrets(key)
-      secrets_json[key.to_s]
-    end
-
-    def push_to_settings(key:, value:)
-      self.settings ||= {}
-      settings[key] = value
-    end
-
-    def get_from_settings(key)
-      (settings || {})[key]
-    end
 
     def webhook_secret=(value)
       push_to_settings(key: 'webhook_secret', value:)
