+++ spec/models/integrations/netsuite_integration_spec.rb
@@ -3,9 +3,7 @@
 require 'rails_helper'
 
 RSpec.describe Integrations::NetsuiteIntegration, type: :model do
-  subject(:netsuite_integration) { build(:netsuite_integration, attributes) }
-
-  let(:attributes) {}
+  subject(:netsuite_integration) { build(:netsuite_integration) }
 
   it { is_expected.to validate_presence_of(:name) }
